break out get_context_user
